motors driver

+++ Project_wsl_2026/Software/systemCore_ver1/ir_seeeduinoxiao/lib/movement_average/movement_average.cpp
@@ -24,7 +24,7 @@ int Movement_average::output()
         }
     }
 
-    double ave = 0.0;
+    float ave = 0.0f;
     if (available_count > 0)
     {
         ave = sum / available_count;
